V-F-2_Aufgabenkatalog: Separate Entites and Dao Interfaces from ui code - update imports
