+++ Integra/app/src/main/res/layout/item_area.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
-        <variable name="area" type="com.dhbw.cas.integra.ui.areas.Area"/>
+        <variable name="area" type="com.dhbw.cas.integra.data.Area"/>
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
